Roll my own prerendering solution
All of the prerendering packages out there that I could find were
abandoned. So, I'm just going to do it myself.

Vue provides createSSRApp and renderToString, which are intended for
server-side rendering, but it's close enough to what I need. I use JSDOM
for rendering, based on the process used by historical prerendering
packages. However, instead of using the resulting DOM verbatim, I
instead replace it with the result of renderToString, which is the only
input that Vue can latch onto without producing hydration errors.

The problem with this approach is that renderToString is supposed to be
run server-side, and I've had to integrate it into the bundle, which
will run client-side. This inflates the bundle size. I will have to find
some way of extracting it out to be imported only during the prerender.
One possibility: execute webpack twice, once to produce the string and
once without renderToString using minimisation and a DefinePlugin flag.
But I can deal with that another time.